Types
fix mongoose model types
make use of .orFail on mongoose query methods
add UserActivityManager
replace deprecated @sendinblue/client with @getbrevo/brevo
new projections class
many bug fixes
cleanup

+++ src/api/controllers/ConnectionController.ts
@@ -3,19 +3,23 @@ import { body, param, query, type ValidationChain } from "express-validator";
 import { StatusCodes } from "http-status-codes";
 import mongoose from "mongoose";
 
-import Follow from "../../models/Follow";
-import FollowRequest, { type IFollowRequest } from "../../models/FollowRequest";
+import Follow, { FollowStatusEnum } from "../../models/Follow";
+import FollowRequest from "../../models/FollowRequest";
 import Notification, { ResourceTypeEnum } from "../../models/Notification";
-import User, { type IUser } from "../../models/User";
+import User from "../../models/User";
 import UserActivity, {
   ActivityResourceTypeEnum,
   ActivityTypeEnum,
 } from "../../models/UserActivity";
-import strings, { dStrings, dynamicMessage } from "../../strings";
+import { dStrings, dynamicMessage } from "../../strings";
+import {
+  getConnectionStatus,
+  type ConnectionStatus,
+} from "../../utilities/connections";
 import { createError, handleInputErrors } from "../../utilities/errorHandlers";
-import UserProjection from "../dto/user/user";
+import UserProjection, { type UserProjectionEssentials } from "../dto/user";
+import { UserActivityManager } from "../services/UserActivityManager";
 import logger from "../services/logger";
-import { addNewFollowingActivity } from "../services/user.activity.service";
 import validate from "./validators";
 
 export const connectionFollowStatusValidation: ValidationChain[] = [
@@ -33,25 +37,11 @@ export async function connectionFollowStatus(
     const { id } = req.params;
     const authUser = req.user!;
 
-    const [followedByUser, followsUser] = await Promise.all([
-      Follow.exists({ user: authUser._id, target: id }),
-      Follow.exists({ user: id, target: authUser._id }),
-    ]);
-
-    const isRequestPending =
-      (await FollowRequest.exists({ user: authUser._id, target: id })) !== null;
+    const connectionStatus = await getConnectionStatus(authUser._id, id);
 
     return res.status(StatusCodes.OK).json({
       success: true,
-      data: {
-        followedByUser: !!followedByUser,
-        followsUser: !!followsUser,
-        isRequestPending,
-
-        // TODO: Remove this after the client is updated
-        isFollowing: !!followedByUser,
-        isFollower: !!followsUser,
-      },
+      data: connectionStatus,
     });
   } catch (error) {
     next(error);
@@ -62,21 +52,6 @@ export const createUserConnectionValidation: ValidationChain[] = [
   param("id").isMongoId(),
 ];
 
-async function logFollowingActivity(
-  authId: mongoose.Types.ObjectId,
-  targetId: string
-) {
-  try {
-    await addNewFollowingActivity(authId, targetId);
-  } catch (e) {
-    logger.error(
-      "Something happened during creation of activity for the following",
-      { error: e }
-    );
-    throw e;
-  }
-}
-
 export async function createUserConnection(
   req: Request,
   res: Response,
@@ -85,27 +60,29 @@ export async function createUserConnection(
   try {
     handleInputErrors(req);
 
-    const { id } = req.params;
     const authUser = req.user!;
 
+    const id = new mongoose.Types.ObjectId(req.params.id);
+
     // Check if the follow relationship already exists
-    const existingFollow = await Follow.findOne({
+    const existingFollow = await Follow.exists({
       user: authUser._id,
       target: id,
     });
     if (existingFollow) {
       logger.debug("You already followed this person");
-      throw createError(strings.follows.alreadyExists, StatusCodes.CONFLICT);
-    }
-
-    const targetUser: IUser | null = await User.findById(id);
-
-    if (!targetUser) {
       throw createError(
-        dynamicMessage(dStrings.notFound, "User"),
-        StatusCodes.NOT_FOUND
+        dynamicMessage(dStrings.alreadyExists, "Follow"),
+        StatusCodes.CONFLICT
       );
     }
+
+    const targetUser = await User.findById(id).orFail(
+      createError(
+        dynamicMessage(dStrings.notFound, "User"),
+        StatusCodes.NOT_FOUND
+      )
+    );
 
     if (targetUser.isPrivate) {
       const existingFollowRequest = await FollowRequest.findOne({
@@ -120,19 +97,27 @@ export async function createUserConnection(
         );
       }
 
-      await FollowRequest.create({
+      const request = await FollowRequest.create({
         user: authUser._id,
         target: id,
       });
 
-      res.status(StatusCodes.NO_CONTENT);
+      res.status(StatusCodes.ACCEPTED).json({
+        success: true,
+        data: request,
+      });
 
       //TODO: Send Notification to Target that they have a follow request
     } else {
       // Create new follow relationship
-      const follow = await Follow.create({ user: authUser._id, target: id });
-      // Log new following activity
-      await logFollowingActivity(authUser._id, id);
+      const follow = await Follow.create({
+        user: authUser._id,
+        target: id,
+      });
+
+      // Create following activity
+      await UserActivityManager.createFollowActivity(authUser, id);
+
       res.status(StatusCodes.CREATED).json({ success: true, data: follow });
     }
   } catch (err) {
@@ -140,60 +125,189 @@ export async function createUserConnection(
   }
 }
 
-export const getPendingConnectionsValidation: ValidationChain[] = [];
+export const getFollowRequestsValidation: ValidationChain[] = [
+  validate.page(query("page").optional()),
+  validate.limit(query("limit").optional(), 10, 50),
+];
 
-export async function getPendingConnections(
+export async function getFollowRequests(
   req: Request,
   res: Response,
   next: NextFunction
 ) {
   try {
     handleInputErrors(req);
+
     const authUser = req.user!;
 
-    const followRequests = await FollowRequest.find({
-      target: authUser._id,
-    }).populate("user", UserProjection.essentials);
+    const limit = parseInt(req.query.limit as string) || 50;
+    const page = parseInt(req.query.page as string) || 1;
+    const skip = (page - 1) * limit;
+
+    const [userFollowRequests, followRequests] = await Promise.all([
+      FollowRequest.find({
+        target: authUser._id,
+      })
+        .sort({ _id: -1 })
+        .skip(skip)
+        .limit(limit)
+        .populate<{
+          user: UserProjectionEssentials & {
+            connectionStatus: ConnectionStatus | null;
+          };
+        }>("user", UserProjection.essentials)
+        .lean(),
+      FollowRequest.find({
+        user: authUser._id,
+      }).lean(),
+    ]);
+
+    // Get follow status for each user
+    const usersObject: Record<string, ConnectionStatus> = {};
+
+    userFollowRequests.forEach((f) => {
+      const userId: string = f.user._id.toString();
+      if (userId in usersObject) {
+        usersObject[userId].followedByStatus = FollowStatusEnum.REQUESTED;
+      } else {
+        usersObject[userId] = {
+          followedByUser: false,
+          followsUser: false,
+          followingStatus: FollowStatusEnum.NOT_FOLLOWING,
+          followedByStatus: FollowStatusEnum.REQUESTED,
+        };
+      }
+    });
+
+    followRequests.forEach((f) => {
+      const targetId: string = f.target.toString();
+      if (targetId in usersObject) {
+        usersObject[targetId].followingStatus = FollowStatusEnum.REQUESTED;
+      } else {
+        usersObject[targetId] = {
+          followedByUser: false,
+          followsUser: false,
+          followingStatus: FollowStatusEnum.REQUESTED,
+          followedByStatus: FollowStatusEnum.NOT_FOLLOWING,
+        };
+      }
+    });
+
+    const usersObjectKeys = Object.keys(usersObject).map(
+      (key) => new mongoose.Types.ObjectId(key)
+    );
+    const followItems = await Follow.find({
+      $or: [
+        {
+          user: authUser._id,
+          target: usersObjectKeys,
+        },
+        {
+          target: authUser._id,
+          user: usersObjectKeys,
+        },
+      ],
+    })
+      .select({
+        target: 1,
+        user: 1,
+      })
+      .lean();
+
+    followItems.forEach((f) => {
+      const userId = f.user.toString();
+      const targetId = f.target.toString();
+      if (authUser._id.equals(userId)) {
+        usersObject[targetId].followedByUser = true;
+        usersObject[targetId].followingStatus = FollowStatusEnum.FOLLOWING;
+      } else {
+        usersObject[userId].followsUser = true;
+        usersObject[userId].followedByStatus = FollowStatusEnum.FOLLOWING;
+      }
+    });
+
+    userFollowRequests.forEach((f) => {
+      f.user.connectionStatus = usersObject[f.user._id.toString()];
+    });
 
     res
-      .status(StatusCodes.CREATED)
-      .json({ success: true, data: followRequests });
+      .status(StatusCodes.OK)
+      .json({ success: true, data: userFollowRequests });
   } catch (error) {
     next(error);
   }
 }
 
-export const acceptConnectionRequestValidation: ValidationChain[] = [
+export const acceptFollowRequestValidation: ValidationChain[] = [
   body("id").isMongoId(),
 ];
 
-export async function acceptConnectionRequest(
+export async function acceptFollowRequest(
   req: Request,
   res: Response,
   next: NextFunction
 ) {
   try {
     handleInputErrors(req);
+
     const { id } = req.body;
     const authUser = req.user!;
 
-    const followRequest: IFollowRequest | null = await FollowRequest.findOne({
+    const followRequest = await FollowRequest.findOne({
       _id: id,
       target: authUser._id,
-    });
-
-    if (!followRequest) {
-      throw createError(strings.followRequest.notFound, StatusCodes.NOT_FOUND);
-    }
+    }).orFail(
+      createError(
+        dynamicMessage(dStrings.notFound, "Follow Request"),
+        StatusCodes.NOT_FOUND
+      )
+    );
 
     const follow = await Follow.create({
       user: followRequest.user,
       target: followRequest.target,
     });
 
-    await followRequest.deleteOne();
+    await Promise.all([
+      UserActivityManager.createFollowActivity(authUser, followRequest.target),
+      followRequest.deleteOne(),
+    ]);
+
     //TODO: Send notification to follow.user that your follow request got accepted
     res.status(StatusCodes.CREATED).json({ success: true, data: follow });
+  } catch (error) {
+    next(error);
+  }
+}
+
+export const rejectFollowRequestValidation: ValidationChain[] = [
+  body("id").isMongoId(),
+];
+
+export async function rejectFollowRequest(
+  req: Request,
+  res: Response,
+  next: NextFunction
+) {
+  try {
+    handleInputErrors(req);
+
+    const { id } = req.body;
+    const authUser = req.user!;
+
+    const followRequest = await FollowRequest.findOne({
+      _id: id,
+      target: authUser._id,
+    }).orFail(
+      createError(
+        dynamicMessage(dStrings.notFound, "Follow Request"),
+        StatusCodes.NOT_FOUND
+      )
+    );
+
+    await followRequest.deleteOne();
+
+    res.sendStatus(StatusCodes.NO_CONTENT);
   } catch (error) {
     next(error);
   }
@@ -234,17 +348,19 @@ export async function deleteUserConnection(
       logger.error("Error while deleting user connection", { error: e });
     }
 
-    try {
-      await Notification.deleteOne({
-        resources: {
-          $elemMatch: { _id: deletedDoc._id, type: ResourceTypeEnum.FOLLOW },
-        },
-      });
-    } catch (e) {
-      logger.error(
-        "Error while deleting notification after deleting user connection",
-        { error: e }
-      );
+    if (deletedDoc) {
+      try {
+        await Notification.deleteOne({
+          resources: {
+            $elemMatch: { _id: deletedDoc._id, type: ResourceTypeEnum.FOLLOW },
+          },
+        });
+      } catch (e) {
+        logger.error(
+          "Error while deleting notification after deleting user connection",
+          { error: e }
+        );
+      }
     }
 
     res.sendStatus(StatusCodes.NO_CONTENT);
@@ -279,7 +395,10 @@ export async function getUserConnections(
 
     const userExists = await User.findById(theUserId).lean();
     if (!userExists) {
-      throw createError(strings.user.notFound, StatusCodes.NOT_FOUND);
+      throw createError(
+        dynamicMessage(dStrings.notFound, "User"),
+        StatusCodes.NOT_FOUND
+      );
     }
 
     const data = await Follow.aggregate([
